Fix on Welsh league Adjusted news sent out for various competitions with group stages Deleted some comments

+++ Brazil/bra_first.cpp
@@ -42,7 +42,7 @@ void bra_first_prom_rel_update(BYTE* _this, int a2) {
 	}
 }
 
-void __declspec(naked) bra_first_prom_rel_update_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_first_prom_rel_update_c()
 {
 	__asm
 	{
@@ -80,14 +80,13 @@ void bra_first_subs(BYTE* _this)
 	comp_data->max_bench = 7;
 	comp_data->max_subs = 3;
 
-	//call vtable +3C which is actually add fixtures function
 	DWORD v1 = *(DWORD*)_this;
 	comp_data->fixtures_table = (DWORD*)(*(int(__thiscall**)(BYTE*, int, BYTE*, BYTE*, DWORD))(v1 + 0x3C))(_this, -1, _this + 0xA9, _this + 0x3A, 0);
 
 	return;
 }
 
-void __declspec(naked) bra_first_subs_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_first_subs_c()
 {
 	__asm
 	{
@@ -272,7 +271,7 @@ DWORD bra_first_fixtures(BYTE* _this, char stage_idx, WORD* num_rounds, WORD* st
 	return 0;
 }
 
-void __declspec(naked) bra_first_fixtures_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_first_fixtures_c()
 {
 	__asm
 	{
@@ -587,7 +586,7 @@ char bra_first_update(BYTE* _this) {
 	return sub_79CEE0((BYTE*)*b74340, (BYTE*)(data->competition_db));
 }
 
-void __declspec(naked) bra_first_update_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_first_update_c()
 {
 	__asm
 	{
@@ -605,10 +604,10 @@ void setup_bra_first()
 	WriteVTablePtr(bra_first_vtable, VTableFixtures, (DWORD)&bra_first_fixtures_c);
 	WriteVTablePtr(bra_first_vtable, VTableEoSUpdate, (DWORD)&bra_first_update_c);
 	WriteVTablePtr(bra_first_vtable, VTablePromRelUpdate, (DWORD)&bra_first_prom_rel_update_c);
-	WriteVTablePtr(bra_first_vtable, VTablePlayoffQual, (DWORD)0x5a8f60);
-	WriteVTablePtr(bra_first_vtable, VTableSetChampion, (DWORD)0x684640);
-	WriteVTablePtr(bra_first_vtable, VTableTableFates, (DWORD)0x686940);
-	WriteVTablePtr(bra_first_vtable, VTableStageNews, (DWORD)0x48c6d0);
-	WriteVTablePtr(bra_first_vtable, VTableReputationCalc, (DWORD)0x48e380);
-	WriteVTablePtr(bra_first_vtable, VTable37, (DWORD)0x68aad0);
+	WriteVTablePtr(bra_first_vtable, VTablePlayoffQual, 0x5a8f60);
+	WriteVTablePtr(bra_first_vtable, VTableSetChampion, 0x684640);
+	WriteVTablePtr(bra_first_vtable, VTableTableFates, 0x686940);
+	WriteVTablePtr(bra_first_vtable, VTableStageNews, 0x48c6d0);
+	WriteVTablePtr(bra_first_vtable, VTableReputationCalc, 0x48e380);
+	WriteVTablePtr(bra_first_vtable, VTable37, 0x68aad0);
 }
