+++ Brazil/bra_fourth.cpp
@@ -16,7 +16,7 @@ int bra_fourth_set_champion(BYTE* _this) {
 	return (*(int(__thiscall**)(BYTE*))(v1 + 0x30))(stage_data_for_history);
 }
 
-void __declspec(naked) bra_fourth_set_champion_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_fourth_set_champion_c()
 {
 	__asm
 	{
@@ -54,14 +54,13 @@ void bra_fourth_subs(BYTE* _this)
 	comp_data->max_bench = 7;
 	comp_data->max_subs = 3;
 
-	//call vtable +3C which is actually add fixtures function
 	DWORD v1 = *(DWORD*)_this;
 	comp_data->fixtures_table = (DWORD*)(*(int(__thiscall**)(BYTE*, int, BYTE*, BYTE*, DWORD))(v1 + 0x3C))(_this, -1, _this + 0xA9, _this + 0x3A, 0);
 
 	return;
 }
 
-void __declspec(naked) bra_fourth_subs_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_fourth_subs_c()
 {
 	__asm
 	{
@@ -156,7 +155,7 @@ void bra_fourth_reputation_calc(BYTE* _this, BYTE* club, char stage, char curren
 	ret[0x75] = ret_max;
 }
 
-void __declspec(naked) bra_fourth_reputation_calc_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_fourth_reputation_calc_c()
 {
 	__asm
 	{
@@ -242,7 +241,7 @@ DWORD bra_fourth_fixtures(BYTE* _this, char stage_idx, WORD* num_rounds, WORD* s
 	return 0;
 }
 
-void __declspec(naked) bra_fourth_fixtures_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_fourth_fixtures_c()
 {
 	__asm
 	{
@@ -298,7 +297,7 @@ void bra_fourth_reputation_setup(BYTE* _this) {
 	}
 }
 
-void __declspec(naked) bra_fourth_reputation_setup_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_fourth_reputation_setup_c()
 {
 	__asm
 	{
@@ -377,7 +376,7 @@ char bra_fourth_update(BYTE* _this) {
 	return sub_79CEE0((BYTE*)*b74340, (BYTE*)(data->competition_db));
 }
 
-void __declspec(naked) bra_fourth_update_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_fourth_update_c()
 {
 	__asm
 	{
@@ -457,7 +456,7 @@ int bra_fourth_set_fates(BYTE* _this, cm3_clubs* club, char fate, char stage, BY
 	return 0;
 }
 
-void __declspec(naked) bra_fourth_set_table_fate()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_fourth_set_table_fate()
 {
 	__asm
 	{
@@ -571,7 +570,7 @@ void bra_fourth_playoffs_c(BYTE* _this) {
 	}
 }
 
-void __declspec(naked) bra_fourth_playoffs_create()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_fourth_playoffs_create()
 {
 	__asm
 	{
@@ -591,7 +590,7 @@ void setup_bra_fourth() {
 	WriteVTablePtr(bra_fourth_vtable, VTableReputationCalc, (DWORD)&bra_fourth_reputation_calc_c);
 	WriteVTablePtr(bra_fourth_vtable, VTableSetChampion, (DWORD)&bra_fourth_set_champion_c);
 	WriteVTablePtr(bra_fourth_vtable, VTableTableFates, (DWORD)&bra_fourth_set_table_fate);
-	WriteVTablePtr(bra_fourth_vtable, VTableStageNews, (DWORD)0x48c6d0);
+	WriteVTablePtr(bra_fourth_vtable, VTableStageNews, 0x48c6d0);
 	WriteVTablePtr(bra_fourth_vtable, VTablePlayoffQual, (DWORD)&bra_fourth_playoffs_create);
-	WriteVTablePtr(bra_fourth_vtable, VTable37, (DWORD)0x68aad0);
+	WriteVTablePtr(bra_fourth_vtable, VTable37, 0x68aad0);
 }
