+++ Brazil/bra_reg_goias.cpp
@@ -15,7 +15,7 @@ int bra_reg_goias_set_champion(BYTE* _this) {
 	return (*(int(__thiscall**)(BYTE*))(v1 + 0x30))(stage_data_for_history);
 }
 
-void __declspec(naked) bra_reg_goias_set_champion_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_reg_goias_set_champion_c()
 {
 	__asm
 	{
@@ -44,7 +44,7 @@ void bra_reg_goias_reputation_calc(BYTE* _this, BYTE* club, char stage, char cur
 	ret[0x75] = ret_max;
 }
 
-void __declspec(naked) bra_reg_goias_reputation_calc_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_reg_goias_reputation_calc_c()
 {
 	__asm
 	{
@@ -87,14 +87,13 @@ void bra_reg_goias_subs(BYTE* _this)
 	comp_data->max_bench = 7;
 	comp_data->max_subs = 3;
 
-	//call vtable +3C which is actually add fixtures function
 	DWORD v1 = *(DWORD*)_this;
 	comp_data->fixtures_table = (DWORD*)(*(int(__thiscall**)(BYTE*, int, BYTE*, BYTE*, DWORD))(v1 + 0x3C))(_this, -1, _this + 0xA9, _this + 0x3A, 0);
 
 	return;
 }
 
-void __declspec(naked) bra_reg_goias_subs_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_reg_goias_subs_c()
 {
 	__asm
 	{
@@ -198,7 +197,7 @@ DWORD bra_reg_goias_fixtures(BYTE* _this, char stage_idx, WORD* num_rounds, WORD
 	return 0;
 }
 
-void __declspec(naked) bra_reg_goias_fixtures_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_reg_goias_fixtures_c()
 {
 	__asm
 	{
@@ -270,7 +269,7 @@ int bra_reg_goias_set_fates(BYTE* _this, cm3_clubs* club, char fate, char stage,
 	return 0;
 }
 
-void __declspec(naked) bra_reg_goias_set_table_fate()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_reg_goias_set_table_fate()
 {
 	__asm
 	{
@@ -321,7 +320,7 @@ char bra_reg_goias_update(BYTE* _this) {
 	return sub_79CEE0((BYTE*)*b74340, (BYTE*)(data->competition_db));
 }
 
-void __declspec(naked) bra_reg_goias_update_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_reg_goias_update_c()
 {
 	__asm
 	{
@@ -371,7 +370,7 @@ void bra_reg_goias_playoffs_c(BYTE* _this) {
 	}
 }
 
-void __declspec(naked) bra_reg_goias_playoffs_create()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_reg_goias_playoffs_create()
 {
 	__asm
 	{
@@ -413,7 +412,7 @@ void bra_reg_goias_init(BYTE* _this, WORD year, cm3_club_comps* comp) {
 
 void setup_bra_reg_goias() {
 	WriteVTablePtr(bra_reg_goias_vtable, VTableSubsRounds, (DWORD)&bra_reg_goias_subs_c);
-	WriteVTablePtr(bra_reg_goias_vtable, VTableReputationSetup, (DWORD)0x68a850);
+	WriteVTablePtr(bra_reg_goias_vtable, VTableReputationSetup, 0x68a850);
 	WriteVTablePtr(bra_reg_goias_vtable, VTableFixtures, (DWORD)&bra_reg_goias_fixtures_c);
 	WriteVTablePtr(bra_reg_goias_vtable, VTableEoSUpdate, (DWORD)&bra_reg_goias_update_c);
 	WriteVTablePtr(bra_reg_goias_vtable, VTableReputationCalc, (DWORD)&bra_reg_goias_reputation_calc_c);
