+++ Brazil/bra_second.cpp
@@ -33,14 +33,13 @@ void bra_second_subs(BYTE* _this)
 	comp_data->max_bench = 7;
 	comp_data->max_subs = 3;
 
-	//call vtable +3C which is actually add fixtures function
 	DWORD v1 = *(DWORD*)_this;
 	comp_data->fixtures_table = (DWORD*)(*(int(__thiscall**)(BYTE*, int, BYTE*, BYTE*, DWORD))(v1 + 0x3C))(_this, -1, _this + 0xA9, _this + 0x3A, 0);
 
 	return;
 }
 
-void __declspec(naked) bra_second_subs_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_second_subs_c()
 {
 	__asm
 	{
@@ -225,7 +224,7 @@ DWORD bra_second_fixtures(BYTE* _this, char stage_idx, WORD* num_rounds, WORD* s
 	return 0;
 }
 
-void __declspec(naked) bra_second_fixtures_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_second_fixtures_c()
 {
 	__asm
 	{
@@ -304,7 +303,7 @@ char bra_second_update(BYTE* _this) {
 	return sub_79CEE0((BYTE*)*b74340, (BYTE*)(data->competition_db));
 }
 
-void __declspec(naked) bra_second_update_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_second_update_c()
 {
 	__asm
 	{
@@ -321,10 +320,10 @@ void setup_bra_second()
 	WriteVTablePtr(bra_second_vtable, VTableSubsRounds, (DWORD)&bra_second_subs_c);
 	WriteVTablePtr(bra_second_vtable, VTableFixtures, (DWORD)&bra_second_fixtures_c);
 	WriteVTablePtr(bra_second_vtable, VTableEoSUpdate, (DWORD)&bra_second_update_c);
-	WriteVTablePtr(bra_second_vtable, VTablePlayoffQual, (DWORD)0x5a8f60);
-	WriteVTablePtr(bra_second_vtable, VTableSetChampion, (DWORD)0x684640);
-	WriteVTablePtr(bra_second_vtable, VTableTableFates, (DWORD)0x686940);
-	WriteVTablePtr(bra_second_vtable, VTableStageNews, (DWORD)0x48c6d0);
-	WriteVTablePtr(bra_second_vtable, VTableReputationCalc, (DWORD)0x48e380);
-	WriteVTablePtr(bra_second_vtable, VTable37, (DWORD)0x68aad0);
+	WriteVTablePtr(bra_second_vtable, VTablePlayoffQual, 0x5a8f60);
+	WriteVTablePtr(bra_second_vtable, VTableSetChampion, 0x684640);
+	WriteVTablePtr(bra_second_vtable, VTableTableFates, 0x686940);
+	WriteVTablePtr(bra_second_vtable, VTableStageNews, 0x48c6d0);
+	WriteVTablePtr(bra_second_vtable, VTableReputationCalc, 0x48e380);
+	WriteVTablePtr(bra_second_vtable, VTable37, 0x68aad0);
 }
