+++ Brazil/bra_third.cpp
@@ -15,7 +15,7 @@ int bra_third_set_champion(BYTE* _this) {
 	return (*(int(__thiscall**)(BYTE*))(v1 + 0x30))(stage_data_for_history);
 }
 
-void __declspec(naked) bra_third_set_champion_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_third_set_champion_c()
 {
 	__asm
 	{
@@ -65,7 +65,7 @@ void bra_third_free(BYTE* _this, BYTE a2) {
 	}
 }
 
-void __declspec(naked) bra_third_free_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_third_free_c()
 {
 	__asm
 	{
@@ -104,14 +104,13 @@ void bra_third_subs(BYTE* _this)
 	comp_data->max_bench = 7;
 	comp_data->max_subs = 3;
 
-	//call vtable +3C which is actually add fixtures function
 	DWORD v1 = *(DWORD*)_this;
 	comp_data->fixtures_table = (DWORD*)(*(int(__thiscall**)(BYTE*, int, BYTE*, BYTE*, DWORD))(v1 + 0x3C))(_this, -1, _this + 0xA9, _this + 0x3A, 0);
 
 	return;
 }
 
-void __declspec(naked) bra_third_subs_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_third_subs_c()
 {
 	__asm
 	{
@@ -263,7 +262,7 @@ DWORD bra_third_fixtures(BYTE* _this, char stage_idx, WORD* num_rounds, WORD* st
 	return 0;
 }
 
-void __declspec(naked) bra_third_fixtures_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_third_fixtures_c()
 {
 	__asm
 	{
@@ -313,7 +312,7 @@ char bra_third_update(BYTE* _this) {
 	return sub_79CEE0((BYTE*)*b74340, (BYTE*)(data->competition_db));
 }
 
-void __declspec(naked) bra_third_update_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_third_update_c()
 {
 	__asm
 	{
@@ -347,7 +346,7 @@ void bra_third_reputation_calc(BYTE* _this, BYTE* club, char stage, char current
 	ret[0x75] = ret_max;
 }
 
-void __declspec(naked) bra_third_reputation_calc_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_third_reputation_calc_c()
 {
 	__asm
 	{
@@ -452,7 +451,7 @@ void bra_third_playoffs_create(BYTE* _this) {
 	}
 }
 
-void __declspec(naked) bra_third_playoffs_create_c()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_third_playoffs_create_c()
 {
 	__asm
 	{
@@ -571,7 +570,7 @@ int bra_third_set_fates(BYTE* _this, cm3_clubs* club, char fate, char stage, BYT
 	return 0;
 }
 
-void __declspec(naked) bra_third_set_table_fate()		// used as a __thiscall -> __cdecl converter
+void __declspec(naked) bra_third_set_table_fate()
 {
 	__asm
 	{
@@ -602,9 +601,9 @@ void bra_third_init(BYTE* _this, WORD year, cm3_club_comps* comp)
 	bra_third_vtable->SetPointer(VTablePlayoffQual, (DWORD)&bra_third_playoffs_create_c);
 	bra_third_vtable->SetPointer(VTableSetChampion, (DWORD)&bra_third_set_champion_c);
 	bra_third_vtable->SetPointer(VTableTableFates, (DWORD)&bra_third_set_table_fate);
-	bra_third_vtable->SetPointer(VTableStageNews, (DWORD)0x48c6d0);
+	bra_third_vtable->SetPointer(VTableStageNews, 0x48c6d0);
 	bra_third_vtable->SetPointer(VTableReputationCalc, (DWORD)&bra_third_reputation_calc_c);
-	bra_third_vtable->SetPointer(VTable37, (DWORD)0x68aad0);
+	bra_third_vtable->SetPointer(VTable37, 0x68aad0);
 	data->year = year;
 	data->rules = 0x7;
 	int loaded = sub_687B10(_this, 1);
